+++ arrow-libs/core/arrow-continuations/build.gradle
@@ -28,5 +28,8 @@ kotlin {
                 implementation project(":arrow-core-test")
             }
         }
+        jsMain {
+
+        }
     }
 }
