+++ arrow-libs/core/arrow-core/src/commonTest/kotlin/arrow/core/NonFatalTest.kt
@@ -2,7 +2,6 @@ package arrow.core
 
 import arrow.core.test.UnitSpec
 import io.kotest.matchers.shouldBe
-import io.kotest.assertions.throwables.shouldThrowAny
 
 class NonFatalTest : UnitSpec() {
   init {
@@ -22,29 +21,6 @@ class NonFatalTest : UnitSpec() {
     "Test nonfatals using Throwable#nonFatalOrThrow" {
       nonFatals.forEach {
         it.nonFatalOrThrow() shouldBe it
-      }
-    }
-
-    val fatals: List<Throwable> =
-      listOf(
-        InterruptedException(),
-        StackOverflowError(),
-        OutOfMemoryError(),
-        LinkageError(),
-        object : VirtualMachineError() {
-        }
-      )
-
-    "Test fatals using #invoke()" {
-      fatals.forEach {
-        NonFatal(it) shouldBe false
-      }
-    }
-    "Test fatals using Throwable#nonFatalOrThrow" {
-      fatals.forEach {
-        shouldThrowAny {
-          it.nonFatalOrThrow()
-        }
       }
     }
   }
