+++ arrow-libs/core/arrow-core/src/jsMain/kotlin/arrow/core/ArrowCoreInternalException.kt
@@ -1,0 +1,7 @@
+package arrow.core
+
+@PublishedApi
+internal actual object ArrowCoreInternalException : RuntimeException(
+  "Arrow-Core internal error. Please let us know and create a ticket at https://github.com/arrow-kt/arrow-core/issues/new/choose",
+  null
+)
