+++ arrow-libs/core/arrow-core/src/jsMain/kotlin/arrow/core/NonFatal.kt
@@ -1,0 +1,10 @@
+package arrow.core
+
+import arrow.continuations.generic.ControlThrowable
+import kotlin.coroutines.cancellation.CancellationException
+
+actual fun NonFatal(t: Throwable): Boolean =
+  when (t) {
+    is ControlThrowable, is CancellationException -> false
+    else -> true
+  }
