+++ arrow-libs/gradle.properties
@@ -30,7 +30,7 @@ KOTLIN_POET_VERSION=1.5.0
 KOTLIN_TEST_VERSION=3.3.3
 KOTLIN_VERSION=1.4.32
 KOTLINX_COLLECTIONS_IMMUTABLE_VERSION=0.1
-KOTLINX_COROUTINES_VERSION=1.4.2
+KOTLINX_COROUTINES_VERSION=1.5.0
 KOTLINX_VALIDATOR_VERSION=0.2.2
 KTLINT_GRADLE_VERSION=9.2.1
 KTLINT_VERSION=0.39.0
