+++ arrow-libs/gradle/subproject-mpp.gradle
@@ -29,6 +29,10 @@ kotlin {
             targetCompatibility = JavaVersion.VERSION_1_8
         }
     }
+    js(IR) {
+        browser()
+        nodejs()
+    }
 }
 
 ktlint {
